Add ability to display quizzes for taking.
Successfully displays all types of questions. There is one JSP
for displaying each type of question, and they (quite sketchily)
directly call the DB in order to get the question they're after.
The layout of the questions is confusing, it needs some CSS help
to not be as ugly. Make some changes to Quiz so it would store
retrieve booleans (for the multiple pages field). This may be
slightly buggy still, watch it in future.